requirements.txt
pandas
requests
python-dotenv
google-generativeai
openpyxl
